verify run_and_check on modal and fix wmma example

+++ src/prompts/hardware/model_ex_wmma.py
@@ -7,12 +7,12 @@ class Model(nn.Module):
         super().__init__()
 
     def forward(self, a, b):
-        # Perform matrix multiplication: a (1,128) x b (128,128) = (1,128)
-        return torch.matmul(a, b)
+        # Perform matrix multiplication: a (128,128) x b (128,128) = (128,128)
+        return torch.matmul(a.to(torch.half), b.to(torch.half)).float()
 
 def get_inputs():
     # Generate input tensors for the model:
-    # 'a' has shape (1, 128) and 'b' has shape (128, 128) so that torch.matmul(a, b) is valid.
+    # 'a' has shape (128, 128) and 'b' has shape (128, 128) so that torch.matmul(a, b) is valid.
     a = torch.randn(128, 128).cuda()
     b = torch.randn(128, 128).cuda()
     return [a, b]
